MAINT split test_testing into three files (#2601)
